// BarcodeScannerAngular/src/app/Interfaces/Sample.ts
export interface Sample {

    SampleID: number; 
    LocationID: number; 
    LocationName: string; 
    Barcode: string; 
    Style: string; 
    Season: string; 
    Sty_Num: string; 
    Sty_Qual: string; 
    BancTecStyle: string; 
    Description: string; 
    Sty_size: string; 
    Colour: string ; 
    Qty: number; 
    LoanQty: number; 
    CostPrice: number; 
    GBP: number; 
    USD: number; 
    EUR: number; 
    AUD: number; 
    Range: string; 
    Comp_Div: string; 
    SellSeason: string; 
    ProductGroup: string; 
    ComplimentaryDescription: string; 
    ComplimentaryCollection: string; 
    Returned: boolean; 
    CIMS_QTY_INDEX: string; 
    STY_RET_TYPE: string 
    IsBadBarcode: boolean; 
    USI: number; 
    RetailPrice: number; 
}